refactor: atom 기준 변화에 따른 menu 수정
