--- src/components/atoms/main/header/AccountLogo.tsx
@@ -1,12 +1,0 @@
-import { AccountLogoImage } from "./AccountLogoStyles";
-
-interface Props {
-  src: string;
-  alt: string;
-}
-
-const AccountLogo = ({ src, alt }: Props) => {
-  return <AccountLogoImage src={src} alt={alt} />;
-};
-
-export default AccountLogo;
