--- src/components/atoms/main/header/AccountLogoStyles.ts
@@ -1,8 +1,0 @@
-import tw from "twin.macro";
-import styled from "styled-components";
-
-export const AccountLogoImage = styled.img`
-  ${tw`
-    w-4
-  `}
-`;
