--- src/components/atoms/main/header/AccountText.tsx
@@ -1,11 +1,0 @@
-import { AccountTextParagraph } from "./AccountTextStyles";
-
-interface Props {
-  text: string;
-}
-
-const AccountText = ({ text }: Props) => {
-  return <AccountTextParagraph>{text}</AccountTextParagraph>;
-};
-
-export default AccountText;
