--- src/components/atoms/main/header/AccountTextStyles.ts
@@ -1,10 +1,0 @@
-import tw from "twin.macro";
-import styled from "styled-components";
-
-export const AccountTextParagraph = styled.p`
-  ${tw`
-    font-bold
-    font-['Inter']
-    text-white
-  `}
-`;
