+++ src/components/molecules/main/header/Menus.tsx
@@ -1,14 +1,14 @@
-import tabMenuItems from "~/components/atoms/main/header/TabMenuItems";
-import Menu from "./Menu";
+import tabMenuItems from "~/components/atoms/main/header/tabMenuItems";
+import Menu from "~/components/atoms/main/header/Menu";
 import { MenusBox } from "./MenusStyles";
 
 const Menus = () => {
   return (
     <MenusBox>
-      <Menu to="/" text="CLUB" tabMenuItems={tabMenuItems.club} />
-      <Menu to="/" text="NEWS" tabMenuItems={tabMenuItems.news} />
-      <Menu to="/" text="TEAM ROOM" tabMenuItems={tabMenuItems.teamroom} />
-      <Menu to="/" text="COMMUNITY" tabMenuItems={tabMenuItems.community} />
+      <Menu text="CLUB" tabMenuItems={tabMenuItems.club} />
+      <Menu text="NEWS" tabMenuItems={tabMenuItems.news} />
+      <Menu text="TEAM ROOM" tabMenuItems={tabMenuItems.teamroom} />
+      <Menu text="COMMUNITY" tabMenuItems={tabMenuItems.community} />
     </MenusBox>
   );
 };
